# docs/openapi/trivida-openapi.yaml
openapi: 3.0.3
info:
  title: Trivida API
  version: 1.0.0
  description: >-
    Backend REST API for Trivida MVP (Node/Express + MongoDB). JWT access 15 min, refresh rotation 7 days.
servers:
  - url: https://api.trivida.example.com
    description: Production
  - url: http://localhost:4000
    description: Local Dev
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CronKey:
      type: apiKey
      in: header
      name: X-Cron-Key
...
  schemas:
...
    User:
      type: object
      required: [_id, email, locale, currency]
      properties:
        _id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        locale: { type: string, enum: [fr, en], default: fr }
        currency: { type: string, default: XAF }
        expoPushTokens:
          type: array
          items: { type: string }
...
paths:
...
  /users:
    patch:
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                locale: { type: string, enum: [fr, en] }
                currency: { type: string }
                settings: { type: object }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/User' } } } } } }
  /users/push-tokens:
    post:
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '201': { description: Saved }
    delete:
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '204': { description: Deleted }
...